+++ lib/src/core/network/network_request.dart
@@ -1,16 +1,22 @@
-import 'package:nearby_city/nearby_city.dart';
+import 'dart:convert';
 
-class NetworkRequest extends NetworkHandler {
-  NetworkRequest() {
-    _remoteHandler = RemoteHandler();
-  }
+import 'package:http/http.dart' as http;
 
-  late RemoteHandler _remoteHandler;
-
-  @override
+class NetworkRequest {
   Future<List<dynamic>> fetchNearbyCity() async {
+    final String fileUrl =
+        'https://raw.githubusercontent.com/jwlim94/nearby_city_data/main/under_20km_final.json';
+
     try {
-      return await _remoteHandler.fetchNearbyCity();
+      var response = await http.get(
+        Uri.parse(fileUrl),
+      );
+
+      if (response.statusCode == 200) {
+        return json.decode(response.body);
+      } else {
+        throw Exception('Unable to fetch nearby city');
+      }
     } catch (_) {
       throw Exception('Unable to fetch nearby city');
     }