+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 0.0.1
 
 - initial release
-- add completed `getNearbyCity` functionality
+- add `getNearbyCityByTown` functionality
+
+## 0.0.2
+
+- add `getNearbyCityByLatLng` functionality